5f5c64e BUG: using InputPixelType to hold label count can overflow.

+++ Modules/Segmentation/LabelVoting/include/itkMultiLabelSTAPLEImageFilter.hxx
@@ -216,7 +216,8 @@ MultiLabelSTAPLEImageFilter< TInputImage, TOutputImage, TWeights >
 ::GenerateData()
 {
   // determine the maximum label in all input images
-  this->m_TotalLabelCount = this->ComputeMaximumInputValue() + 1;
+  this->m_TotalLabelCount =
+    static_cast<size_t>(this->ComputeMaximumInputValue()) + 1;
 
   if ( ! this->m_HasLabelForUndecidedPixels )
     {
